+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## x.x.x (xxxx-xx-xx)
 - **Removed `hello` command**: the terminal window isn't very high so every space counts, so I removed this test command, even though it's a historical command.
+- **Added `neofetch` command**: a toy command to just flex that you use ParvaOS btw!
 
 ## 0.0.4 (2025-05-18)
 - **Processes**: added the ability to create multiple processes inside the OS